ADD: Add Finalizized BESC v1.1.1 code

+++ firmware/Recoil-Motor-Controller-B-G431B-ESC1/Core/Inc/motor.h
@@ -21,10 +21,10 @@
 typedef struct {
   uint32_t pole_pairs;
   uint32_t kv_rating;
-  int8_t phase_order;
-  float phase_resistance;
-  float phase_inductance;
-  float max_calibration_current;
+  int32_t  phase_order;
+  float    phase_resistance;
+  float    phase_inductance;
+  float    max_calibration_current;
 } Motor;
